Fix time functions to always use double

+++ examples/02-cube/cube.cpp
@@ -114,7 +114,7 @@ void Run()
 
         const float ratio = static_cast<Rndr::real>(window.GetWidth())
                             / static_cast<Rndr::real>(window.GetHeight());
-        const float angle = std::fmod(10 * Rndr::GetSystemTime(), 360.0f);
+        const float angle = static_cast<float>(std::fmod(10 * Rndr::GetSystemTime(), 360.0));
         const math::Transform t = math::Translate(math::Vector3(0.0f, 0.0f, -3.5f))
                                   * math::Rotate(angle, math::Vector3(1.0f, 1.0f, 1.0f));
         const math::Matrix4x4 p = math::Perspective_RH_N1(45.0f, ratio, 0.1f, 1000.0f);
